seperate out db
