+++ ingest/config/config.go
@@ -18,12 +18,3 @@ var DatabasePort = os.Getenv("DB_PORT")
 var DatabaseName = os.Getenv("DB_NAME")
 var DatabaseUser = os.Getenv("DB_USER")
 var DatabasePassword = os.Getenv("DB_PASSWORD")
-
-var Banner = `
-███╗   ███╗ █████╗ ██████╗  █████╗  ██████╗██╗  ██╗███████╗    ██╗███╗   ██╗ ██████╗ ███████╗███████╗████████╗
-████╗ ████║██╔══██╗██╔══██╗██╔══██╗██╔════╝██║  ██║██╔════╝    ██║████╗  ██║██╔════╝ ██╔════╝██╔════╝╚══██╔══╝
-██╔████╔██║███████║██████╔╝███████║██║     ███████║█████╗      ██║██╔██╗ ██║██║  ███╗█████╗  ███████╗   ██║   
-██║╚██╔╝██║██╔══██║██╔═══╝ ██╔══██║██║     ██╔══██║██╔══╝      ██║██║╚██╗██║██║   ██║██╔══╝  ╚════██║   ██║   
-██║ ╚═╝ ██║██║  ██║██║     ██║  ██║╚██████╗██║  ██║███████╗    ██║██║ ╚████║╚██████╔╝███████╗███████║   ██║   
-╚═╝     ╚═╝╚═╝  ╚═╝╚═╝     ╚═╝  ╚═╝ ╚═════╝╚═╝  ╚═╝╚══════╝    ╚═╝╚═╝  ╚═══╝ ╚═════╝ ╚══════╝╚══════╝   ╚═╝
-`
