+++ ingest/database/db.go
@@ -1,4 +1,4 @@
-package service
+package database
 
 import (
 	"fmt"
