+++ ingest/service/vehicle_service.go
@@ -1,27 +1,28 @@
 package service
 
 import (
+	"ingest/database"
 	"ingest/model"
 	"strings"
 )
 
 func GetAllVehicles() []model.Vehicle {
 	var vehicles []model.Vehicle
-	DB.Find(&vehicles)
+	database.DB.Find(&vehicles)
 	return vehicles
 }
 
 func GetVehicleByID(id string) model.Vehicle {
 	var vehicle model.Vehicle
-	DB.Where("id = ?", id).First(&vehicle)
+	database.DB.Where("id = ?", id).First(&vehicle)
 	return vehicle
 }
 
 func CreateVehicle(vehicle model.Vehicle) error {
-	result := DB.Create(&vehicle)
+	result := database.DB.Create(&vehicle)
 	if result.Error != nil {
 		if strings.Contains(result.Error.Error(), "Duplicate entry") {
-			result = DB.Where("id = ?", vehicle.ID).Updates(&vehicle)
+			result = database.DB.Where("id = ?", vehicle.ID).Updates(&vehicle)
 			if result.Error != nil {
 				return result.Error
 			}
